+++ SlicerCART/src/scripts/InteractingClasses.py
@@ -33,6 +33,8 @@ class SlicerCARTConfigurationSetupWindow(qt.QWidget):
                         CONFIG_FILE_PATH)
 
         self.config_yaml  = ConfigPath.open_project_config_file(self)
+        # self.config_yaml  = ConfigPath.get_config_values(self)
+
 
         self.set_default_values()
 
@@ -406,6 +408,12 @@ class SlicerCARTConfigurationSetupWindow(qt.QWidget):
 
     @enter_function
     def set_default_values(self):
+
+        ConfigPath.write_config_file(self)
+
+        self.config_yaml = ConfigPath.get_config_values(self)
+
+
         self.segmentation_selected = (
             self.config_yaml)['is_segmentation_requested']
         self.classification_selected = (
@@ -446,6 +454,8 @@ class SlicerCARTConfigurationSetupWindow(qt.QWidget):
         self.config_yaml['KEYBOARD_SHORTCUTS'][5]['shortcut']
         self.interpolate_ks_selected = \
         self.config_yaml['KEYBOARD_SHORTCUTS'][6]['shortcut']
+
+
 
     def classification_checkbox_state_changed(self):
         self.classification_selected = self.classification_task_checkbox.isChecked()
@@ -607,7 +617,10 @@ class SlicerCARTConfigurationSetupWindow(qt.QWidget):
 
         ConfigPath.write_config_file(self)
 
+        self.segmenter.config_yaml = ConfigPath.get_config_values(self)
+
         self.segmenter.setup_configuration()
+
 
         self.close()
