+++ SlicerCART/src/utils/global_variables.py
@@ -45,6 +45,15 @@ print('SCRIPT_PATH used in global variables:,', SCRIPT_PATH)
 CONFIG_FILE_PATH = os.path.join(SCRIPT_PATH, 'configuration_config.yml')
 print('CONFIG_FILE_PATH used in global variables:,', CONFIG_FILE_PATH)
 
+with open(CONFIG_FILE_PATH, 'r') as file:
+    content = yaml.safe_load(file)
+
+INITIAL_CONFIG_FILE = content
+
+print(' \n \n \n 666666666 content bg', content)
+
+print('len content initial', len(content))
+
 # This variable enables/disables easier debug mode (with print) in the python
 # console (e.g. Debut.set_debug(self, True) or Debut.set_debug(self, False)
 # ENABLE_DEBUG = True
